+++ hutool-core/src/test/java/cn/hutool/core/math/CalculatorTest.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.math;
 
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 
 public class CalculatorTest {
@@ -26,6 +27,14 @@ public class CalculatorTest {
 	@Test
 	public void conversationTest4(){
 		final double conversion = Calculator.conversion("(88*66/23)%26+45%9");
+		Assert.assertEquals((88D * 66 / 23) % 26, conversion, 2);
+	}
+
+	@Test
+	@Ignore
+	public void conversationTest5(){
+		// https://github.com/dromara/hutool/issues/1984
+		final double conversion = Calculator.conversion("((1/1) / (1/1) -1) * 100");
 		Assert.assertEquals((88D * 66 / 23) % 26, conversion, 2);
 	}
 }