+++ hutool-core/src/test/java/cn/hutool/core/text/CharSequenceUtilTest.java
@@ -1,5 +1,6 @@
 package cn.hutool.core.text;
 
+import cn.hutool.core.util.CharsetUtil;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -64,5 +65,17 @@ public class CharSequenceUtilTest {
 
 		index = CharSequenceUtil.indexOf("abc123", 'b', 0, 3);
 		Assert.assertEquals(1, index);
+	}
+
+	@Test
+	public void subPreGbkTest(){
+		// https://gitee.com/dromara/hutool/issues/I4JO2E
+		String s = "华硕K42Intel酷睿i31代2G以下独立显卡不含机械硬盘固态硬盘120GB-192GB4GB-6GB";
+
+		String v = CharSequenceUtil.subPreGbk(s, 40, false);
+		Assert.assertEquals(39, v.getBytes(CharsetUtil.CHARSET_GBK).length);
+
+		v = CharSequenceUtil.subPreGbk(s, 40, true);
+		Assert.assertEquals(41, v.getBytes(CharsetUtil.CHARSET_GBK).length);
 	}
 }